Backend/audio-processor/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-multipart==0.0.6
requests==2.28.2
numpy==1.24.2
pydantic==1.10.7
python-dotenv==1.0.0